[Chore #163772707] Restructure the folder structure and models data containers
