+++ app/__init__.py
@@ -2,7 +2,8 @@
 from flask import Flask
 from flask_cors import CORS
 
-from app.views import version1, V1
+from app.V1.offices.views import V1
+from app.V1.parties.views import version1
 from config import DevelopmentConfig
 
 app = Flask(__name__)
@@ -11,4 +12,3 @@ app.config.from_object(DevelopmentConfig)
 app.register_blueprint(version1)
 app.register_blueprint(V1)
 
-from . import views, models
