+++ tests/base_test_case.py
@@ -1,7 +1,8 @@
 from unittest import TestCase
 
 from app import app
-from app.models import MOCK_DATABASE
+from app.V1.offices.models import OFFICE_MOCK_DATABASE
+from app.V1.parties.models import MOCK_DATABASE
 
 
 class BaseTestCase(TestCase):
@@ -15,5 +16,6 @@ class BaseTestCase(TestCase):
 
     def tearDown(self):
         """Method to delete data from data structures"""
+
         MOCK_DATABASE['parties'].clear()
-        MOCK_DATABASE['offices'].clear()
+        OFFICE_MOCK_DATABASE['offices'].clear()
