+++ tests/test_political_parties.py
@@ -11,13 +11,13 @@ class TestParties(BaseTestCase):
     def test_get_parties(self):
         """ Test get all political parties"""
         with self.client:
-            response = self.client.get('/api/v1/parties')
+            response = self.client.get('api/v1/parties',headers={'Content-Type': 'application' '/json'})
             self.assertEqual(response._status_code, 200)
 
     def test_create_party(self):
         """Test to create a political party"""
         with self.client:
-            response = self.client.post('api/v1/parties', data=json.dumps(party_to_post),
+            response = self.client.post('/api/v1/parties', data=json.dumps(party_to_post),
                                         headers={'Content-Type': 'application' '/json'})
             self.assertEqual(response._status_code, 201)
             response = self.client.post('api/v1/parties', data=json.dumps(party1_to_post),
